downsize gamestate fields4

+++ pgx/_src/games/chess.py
@@ -297,7 +297,9 @@ def _apply_move(state: GameState, a: Action) -> GameState:
         )
     )
     is_en_passant = (piece == PAWN) & (jnp.abs(a.to - a.from_) == 2)
-    state = state._replace(en_passant=lax.select(is_en_passant, (a.to + a.from_) // 2, -1))
+    state = state._replace(
+      en_passant=jnp.int8(lax.select(is_en_passant, (a.to + a.from_) // 2, -1))
+    )
     # update counters
     captured = (_pieces_at(state.board, a.to) < 0) | is_en_passant
     state = state._replace(
